Early implementation of generating character attributes

+++ src/main/java/pl/coderslab/dogmeat/role/repository/RoleRepository.java
@@ -3,6 +3,6 @@ package pl.coderslab.dogmeat.role.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.dogmeat.role.entity.Role;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Role findByName(String name);
 }
